+++ CHANGES.rst
@@ -14,6 +14,9 @@ New features
 
 Changes
 -------
+- The :func: `tabular_pipeline` uses a :class:`SquashingScaler` instead of a
+  :class:`StandardScaler` for centering and standard scaling numerical features in the.
+  :pr:`1644` by :user:`Simon Dierickx <dierickxsimon>`
 
 Bugfixes
 --------